+++ packages/tidier-vscode/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to the Tidies is documented here.
 
+### 1.1.0
+
+Fixes issues with not being able to rename files on MacOS
+
+- Adds the setting `tidier.renameOverwrite.enabled` which is set to 'auto' by default, 
+  which means it's enabled on MacOS and Windows by default, and disabled on Linux.
+
 ### 1.0.0
 
 Initial release, featuring: